Fixed a number of relative imports and renamed some "utilities" files to match the parent module.
