+++ numbertools/__init__.py
@@ -1,1 +1,1 @@
-from .utilities import *
+from .numbertools import *
