+++ systemtools/_terminal.py
@@ -1,7 +1,7 @@
 import os
 import shlex
 import subprocess
-import timetools
+from pytools import timetools
