+++ tabletools/_table_methods.py
@@ -1,8 +1,11 @@
-import numbertools
-from tabletools import Table
+print("name: ", __name__)
+
+from pytools import numbertools
+
 import os
 import pandas
 import csv
+from ._pandas_table import PandasTable as Table
 
 def getTableType(io, **kwargs):
 	""" Determines what the general layout of the
