+++ timetools/__init__.py
@@ -1,4 +1,4 @@
 from ._duration import Duration 
 from ._timestamp import Timestamp 
 from ._timer import Timer 
-from .utilities import *
+from .time_methods import *
