+++ timetools/time_methods.py
@@ -1,5 +1,5 @@
 
-from .timestamp import Timestamp
+from ._timestamp import Timestamp
 
 def elapsed(loop_number, loop_block, total_loops, timer):
 	""" Prints a line indicating the elapsed progress of the loop described
